Started types.ts file
